aipoweredlms/requirements.txt
python-dotenv==1.0.1
langchain==0.3.21
langchain-community==0.3.20
langchain-chroma==0.2.2
langchain-google-genai==2.0.10
langchain-core==0.3.46
chromadb==0.6.3
PyPDF2==3.0.1
tiktoken==0.5.2
google-generativeai==0.8.4
boto3